+++ module-app/src/main/java/com/example/moduleapp/repository/IRxProductRepository.java
@@ -7,6 +7,6 @@ import io.reactivex.rxjava3.core.Single;
 import java.util.Collection;
 import java.util.Map;
 
-public interface IRxProductRepository extends IRxJooqRepository<Product, Long> {
+public interface IRxProductRepository extends IRxJooqRepository<Product, Integer> {
     Single<Map<Integer, Integer>> getNumberOfPaid(Collection<Integer> productIds);
 }
