+++ module-app/src/main/java/com/example/moduleapp/service/ProductService.java
@@ -1,18 +1,20 @@
 package com.example.moduleapp.service;
 
 import com.example.common.data.request.PageRequest;
-import com.example.common.data.response.PageResponse;
 import com.example.moduleapp.data.request.ProductRequest;
 import com.example.moduleapp.data.response.ProductResponse;
-import com.example.moduleapp.model.tables.pojos.Product;
 import io.reactivex.rxjava3.core.Single;
 
 import java.util.List;
 
 public interface ProductService {
     Single<String> create(ProductRequest productRequest);
-    Single<String> update(Long id,ProductRequest productRequest);
-    Single<String> delete(Long id);
+
+    Single<String> update(Integer id, ProductRequest productRequest);
+
+    Single<String> delete(Integer id);
+
     Single<List<ProductResponse>> findAll(PageRequest pageRequest);
-    Single<String> addToCart(Long id);
+
+    Single<String> addToCart(Integer id);
 }
